% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCorrect.R
\name{basicPF}
\alias{basicPF}
\title{Basic fitting function for stress-strain curve.}
\usage{
basicPF(x, y, sig)
}
\arguments{
\item{x}{Numeric vector as independent variable.}

\item{y}{Numeric vector as dependent variable.}

\item{sig}{Signif to control masking coefficients.}
}
\value{
A list contains independent variable and fitted dependent variable over the maximum value.
}
\description{
Basic fitting function for stress-strain curve.
}
\examples{
x <- TPMdata[,1]
y <- TPMdata[,2]
m <- basicPF(x, y, 0.99)

}
\keyword{internal}
