% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCorrect.R
\name{chrgen2}
\alias{chrgen2}
\title{Internal function for basicPF}
\usage{
chrgen2(x = NULL, order)
}
\arguments{
\item{x}{Any R object.}

\item{order}{Maximum order for linear model.}
}
\value{
Expression for function lm.
}
\description{
Internal function for basicPF
}
\examples{
x <- c(1:7)
chrgen2(x, 5)
}
\keyword{internal}
