% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TCorrect.R
\name{Dvec}
\alias{Dvec}
\title{Intergal area function.}
\usage{
Dvec(x, y)
}
\arguments{
\item{x}{Numeric vector as independent variable.}

\item{y}{Numeric vector as dependent variable.}
}
\value{
A x-dependent function for intergal area.
}
\description{
Intergal area function.
}
\examples{
x <- TPMdata[,1]
y <- TPMdata[,2]
Dvec(x, y)
}
\keyword{internal}
