% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPMplt.R
\name{surfacebld}
\alias{surfacebld}
\title{Build matrix for 3D plots}
\usage{
surfacebld(x, grp = c("eta", "xi"))
}
\arguments{
\item{x}{An "PLTbuilder" object.}

\item{grp}{"eta" or "xi" to determine which group is extracted as the subset.}
}
\value{
A matrix for 3d surface plot.
}
\description{
Build matrix for 3D plots
}
\examples{
epstable <- epsExtract(TPMdata, 0.7, 2, 3)
DMM <- DMMprocess(epstable)
PLTbd <- SVRModel(DMM)
surfacebld(PLTbd, "eta")
}
\keyword{internal}
