% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPMdata.R
\docType{data}
\name{TPMdata}
\alias{TPMdata}
\title{A test data structurized column names.}
\format{
A data frame with 3939 rows and 56 columns
}
\usage{
TPMdata
}
\description{
A test data with 56 different columns.
}
\details{
* `data type`: numeric values, for Strain or Stress.

* `temperature`: from 900 to 1200 celsius, with 50 degrees interval.

* `strain rate`: from slow to speedy: 0.001, 0.01, 0.1, 1.

* `working rate`: 60 percent only.
}
\keyword{internal}
