% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrePLT.R
\name{xitidy}
\alias{xitidy}
\title{Tidy xi table for further process}
\usage{
xitidy(x)
}
\arguments{
\item{x}{A calculation result returned from the function \code{\link[TPMplt:DMMprocess]{DMMprocess}}.}
}
\value{
A matrix with melted temperatures, logarithm strain rates and values of xi values in respective columns.
}
\description{
Tidy the table of xi values, with different logarithm strain rates and temperatures for further
analysis, based on the given dynamic material model.
}
\examples{
epstable <- epsExtract(TPMdata, 0.7, 2, 3)
DMM <- DMMprocess(epstable)
xiM <- xitidy(DMM)
xiM
}
\seealso{
\code{\link[TPMplt:etatidy]{etatidy}}, \code{\link[TPMplt:DMMprocess]{DMMprocess}}
}
\keyword{internal}
