% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMZtester.API.R
\name{API4TMZ}
\alias{API4TMZ}
\title{Internal functions}
\usage{
API4TMZ(Cdl, wd = getwd(), ftype = ".csv")
}
\arguments{
\item{Cdl}{An handmade double list to determine selected conditions.}

\item{wd}{Work directory. Default setting is \code{\link[base:getwd]{getwd()}}.}

\item{ftype}{File type to be read. Defaust setting is ".csv".}
}
\value{
A matrix summary table for all input files.
}
\description{
Internal functions
}
\examples{
\dontrun{
variable1 <- c("factor11", "factor12", "factor13")
variable2 <- c("factor21", "factor22")
variable3 <- c("factor31", "factor32", "factor33", "factor34")
conditions <- list(variable1, variable2, variable3)
SummaryTable <- API4TMZ(conditions, "/Your_Working_Directory/")
SummaryTable
}
}
\keyword{internal}
