% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMMprocess.R
\name{DMMprocess}
\alias{DMMprocess}
\title{Dynamic material modeling from strain rate temperature table}
\usage{
DMMprocess(x, consfuncPRT = FALSE, lgbase = exp(1), rnd = 2)
}
\arguments{
\item{x}{A strain rate-temperature table, returned from \code{\link[TPMplt:epsExtract]{epsExtract}}.}

\item{consfuncPRT}{A boolean value to control result printing for constrcutive function. The
default value uses FALSE.}

\item{lgbase}{A numeric value to determine the logarithm base in calculation. The default
value is exp(1).}

\item{rnd}{An integer to specify the digit accuracy. Default setting is 0.01.}
}
\value{
Serial material constants, constructive function, eta table and xi table through
dynamic material modeling developed by Gegel and Prasad.
}
\description{
Dynamic material modeling from strain rate-temperature table returned from
the function \code{\link[TPMplt:epsExtract]{epsExtract}}. And calculating the constructive
function.
}
\examples{
dl2vbt(chrvec2dl(colnames(TPMdata)))
epstable <- epsExtract(TPMdata, 0.7, 2, 3)
DMM <- DMMprocess(epstable)
DMM
}
\seealso{
\code{\link[VBTree:VBTree-package]{VBTree}}, \code{\link[TPMplt:epsExtract]{epsExtract}}
}
\keyword{"DMMprocess"}
\keyword{"DMMresult"}
\keyword{"epsExtract"}
