% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lyIDdetector.R
\name{lyIDdetector}
\alias{lyIDdetector}
\title{Internal functions}
\usage{
lyIDdetector(data, patterns = "[Ss][Tt][Rr]")
}
\arguments{
\item{data}{Input data frame with the style of \code{\link[VBTree:VBTree-package]{VBTree}}.}

\item{patterns}{Argument to determine layer of Strain and Stress. The default pattern uses "[Ss][Tt][Rr]".}
}
\value{
A list consists of the layer for Strain and Stress, the levels for Strain and Stress respectively in this layer.
}
\description{
Function for detecting the locations for Strain and Stress in data frame. It is an key component for automatic
completion in the function of \code{\link[TPMplt:epsExtract]{epsExtract}}.
}
\examples{
require(VBTree)
chrvec2dl(colnames(TPMdata))
lyIDdetector(TPMdata)
}
\seealso{
\code{\link[VBTree:VBTree-package]{VBTree}}, \code{\link[TPMplt:epsExtract]{TPMplt}}
}
\keyword{internal}
