% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLTblder.R
\name{SVRModel}
\alias{SVRModel}
\title{Build support vector regression result}
\usage{
SVRModel(x, seqby = 80)
}
\arguments{
\item{x}{A calculation result returned from the function \code{\link[TPMplt:DMMprocess]{DMMprocess}}.}

\item{seqby}{A numeric value to specify the grid density. The default value used is 80, namely the
default mesh used 80*80 for original plot.}
}
\value{
A data frame including continuous values for eta and xi, calculated based on the discrete
tables for eta and xi calculated through \code{\link[TPMplt:DMMprocess]{DMMprocess}}.
}
\description{
Return a continuous values for eta and xi, based on prediction built by support vector
regression model. The kernel function is radial basis.
}
\examples{
epstable <- epsExtract(TPMdata, 0.7, 2, 3)
DMM <- DMMprocess(epstable)
PLTbd <- SVRModel(DMM)
PLTbd
}
\seealso{
\code{\link[TPMplt:DMMprocess]{DMMprocess}},
}
\keyword{"PLTbuilder"}
\keyword{"SVRModel"}
