% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMZtester.API.R
\name{TMZdatainput}
\alias{TMZdatainput}
\title{Internal functions}
\usage{
TMZdatainput(makeidx = FALSE, ...)
}
\arguments{
\item{makeidx}{Boolean to control the index column for summary table. Default setting is FALSE.}

\item{...}{Arguments to be past to \code{\link[TPMplt:API4TMZ]{API4TMZ}}.}
}
\value{
A data frame summary table for all input files.
}
\description{
Internal functions
}
\examples{
\dontrun{
variable1 <- c("factor11", "factor12", "factor13")
variable2 <- c("factor21", "factor22")
variable3 <- c("factor31", "factor32", "factor33", "factor34")
conditions <- list(variable1, variable2, variable3)
SummaryTable <- TMZdatainput(Cdl=conditions, wd="/Your_Working_Directory/")
SummaryTable
}
}
\keyword{internal}
