% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPMplt.R
\name{TPM3dplt}
\alias{TPM3dplt}
\title{Plot 3d thermal process maps}
\usage{
TPM3dplt(x, dvs = 5)
}
\arguments{
\item{x}{Regression results from modeling functions such as \code{\link[TPMplt:SVRModel]{SVRModel}}.}

\item{dvs}{A positive integer to set the divisions for all labels in two surface 3d plots. The default value is 5.}
}
\value{
Two 3d surface plots: the left one denotes power dissipation efficiency factor eta, while the right one is for rheological stability
xi. A zero plane for xi value is added in the right plots for determining unstable region.
}
\description{
Return a 3d thermal process result consisted of 3d surfaces for power dissipation efficiency eta and rheological stability
xi respectively.
}
\examples{
epstable <- epsExtract(TPMdata, 0.7, 2, 3)
DMM <- DMMprocess(epstable)
PLTbd <- SVRModel(DMM)
TPM3dplt(PLTbd)
}
\keyword{"PLTbuilder"}
\keyword{"TPM3dplt"}
