\name{TPMdata}
\alias{TPMdata}
\docType{data}
\title{
A test data structurized column names.
}
\description{
A test data with 56 different columns.
}
\usage{data("TPMdata")}
\details{
A test data structurized column names, with two data type "Strain" and "Stress", 7 different temperatures, 4 kinds
strain rates and one level of compression rate.
}
\examples{
TPMdata
}
\keyword{TPMdata}
